requirements.txt
pandas
evernote3
flask
requests
spacy
sqlalchemy
flask-sqlalchemy
